Lease Generator/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pypdf
SQLAlchemy
Flask-SQLAlchemy
werkzeug
reportlab
